// frontend/src/types/Report.ts
export interface Report {
    id: number;
    complaintType: string;
    descriptorType: string;
    agencyName: string;
    locationType: string;
    incidentAddress: string;
    incidentZip: string;
    addressType: string;
    city: string;
    status: string;
    createdDate: string;
    closedDate: string;
    communityBoard: string;
    borough: string;
    openDataChannelType: string;
    latitude: number;
    longitude: number;
}